refactor(image): extract literal tar link writer
Move byte-preserving GNU long-link emission behind an archive-local builder extension so Docker export uses a focused, reusable API without changing behavior.
