+++ crates/image/lib/archive/mod.rs
@@ -1,6 +1,7 @@
 //! Docker archive load/save support.
 
 mod docker;
+mod tar_ext;
 
 //--------------------------------------------------------------------------------------------------
 // Re-Exports
